requirements.txt
pandas>=1.5
pyarrow>=10.0
fastparquet>=0.8.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0